backend/requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6  # Required for file uploads

# LangChain & LangGraph (MANDATORY for state machine)
# Using newer versions for compatibility with Python 3.11+
langchain>=0.3.0
langchain-openai>=0.2.0  # Optional - for OpenAI models
langchain-anthropic>=0.2.0  # Optional - for Claude
langchain-groq>=0.2.0  # For Groq (fast inference)
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Vector store & embeddings
faiss-cpu==1.13.2  # Latest stable version (use faiss-gpu if you have CUDA)
sentence-transformers==2.3.1

# Document processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3

# Utilities
httpx==0.26.0
tenacity==8.2.3  # Retry logic for API calls
loguru==0.7.2  # Better logging

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Optional: PostgreSQL + pgvector (for production)
# psycopg2-binary==2.9.9
# pgvector==0.2.4

